pyproject.toml
[project]
name = "refiner-agent-vertex"
version = "0.1.0"
description = "A sample agent for Vertex AI deployment using the ADK"
authors = [
    {name = "Your Name", email = "your.email@example.com"} # TODO: Update with your details
]
readme = "README.md"
requires-python = ">=3.12" # Matches the example, adjust if your agent needs a different version
license = "Apache License 2.0" # Or your preferred license

[tool.poetry]
packages = [
    { include = "fastapi_backend" }
]

[tool.poetry.dependencies]
aiosqlite = "^0.20.0"
python = ">=3.12"
requests = "^2.31.0" # General HTTP requests
pydantic = "^2.7.0"  # For data validation (you have schemas.py)
python-dotenv = "^1.0.0" # For .env file handling
google-cloud-aiplatform = {extras = ["adk", "agent-engines"], version = "^1.97.0"} # Core ADK and Vertex AI
google-adk = "^1.3.0" # Explicitly pinned ADK version
absl-py = "^2.1.0" # For application-level flags and logging
cloudpickle = "^3.0.0" # For serializing Python objects
aiohttp = "^3.9.0" # For async HTTP communication with Cloud Run agents

# FastAPI dependencies are managed by google-cloud-aiplatform[adk]

# Add any other specific dependencies your sample_agent needs here.
# For example, if your tools.py or subagents use other libraries.
firebase-admin = "^6.8.0"
litellm = "^1.72.0"
deprecated = "^1.2.18"

[tool.poetry.scripts]
# Development commands
fast-local = "fastapi_backend.main:main"       # Run FastAPI with local agent
fast-remote = "fastapi_backend.main:main"      # Run FastAPI with remote agent



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
